fix: add get-mempool-state command
